pyproject.toml
[tool.poetry]
name = "easy_telegram"
version = "0.1.0"
description = "An easy to setup telegram bot lib"
authors = ["Moritz Richter <moritz.richter@ymail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "^13.1"
python-dotenv = "^0.15.0"
SQLAlchemy = "^1.3.20"
APScheduler = "^3.6.3"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"
mypy = "^0.790"
pylint = "^2.6.0"
sqlalchemy-stubs = "^0.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
